refactor: split scene-syncd API routes
