+++ rust/scene-syncd/src/api/mod.rs
@@ -1,1 +1,7 @@
-pub mod routes;
+pub mod common;
+pub mod layout_routes;
+pub mod pie_routes;
+pub mod procedural_routes;
+pub mod scene_routes;
+pub mod semantic_routes;
+pub mod sync_routes;
