+++ rust/scene-syncd/src/main.rs
@@ -1,8 +1,7 @@
 use axum::http::header;
 use axum::http::Method;
-use axum::routing::{get, post};
 use axum::Router;
-use scene_syncd::api::routes::AppState;
+use scene_syncd::api::common::AppState;
 use scene_syncd::db::connect::connect_surreal;
 use scene_syncd::db::SurrealSceneRepository;
 use scene_syncd::unreal::client::UnrealClient;
@@ -51,200 +50,12 @@ async fn main() -> anyhow::Result<()> {
     };
 
     let app = Router::new()
-        .route("/health", get(scene_syncd::api::routes::health))
-        .route(
-            "/scenes/create",
-            post(scene_syncd::api::routes::create_scene),
-        )
-        .route("/scenes/list", post(scene_syncd::api::routes::list_scenes))
-        .route(
-            "/objects/upsert",
-            post(scene_syncd::api::routes::upsert_object),
-        )
-        .route(
-            "/objects/bulk-upsert",
-            post(scene_syncd::api::routes::bulk_upsert_objects),
-        )
-        .route(
-            "/objects/list",
-            post(scene_syncd::api::routes::list_objects),
-        )
-        .route(
-            "/objects/delete",
-            post(scene_syncd::api::routes::delete_object),
-        )
-        .route(
-            "/groups/create",
-            post(scene_syncd::api::routes::create_group),
-        )
-        .route("/groups/list", post(scene_syncd::api::routes::list_groups))
-        .route(
-            "/generator-runs/create",
-            post(scene_syncd::api::routes::create_generator_run),
-        )
-        .route(
-            "/generator-runs/{run_id}",
-            get(scene_syncd::api::routes::get_generator_run),
-        )
-        .route(
-            "/snapshots/create",
-            post(scene_syncd::api::routes::create_snapshot),
-        )
-        .route(
-            "/snapshots/list",
-            post(scene_syncd::api::routes::list_snapshots),
-        )
-        .route(
-            "/snapshots/restore",
-            post(scene_syncd::api::routes::restore_snapshot),
-        )
-        .route(
-            "/sync/plan",
-            post(scene_syncd::api::routes::plan_sync_route),
-        )
-        .route(
-            "/sync/apply",
-            post(scene_syncd::api::routes::apply_sync_route),
-        )
-        // P3: Semantic routes
-        .route(
-            "/entities/bulk-upsert",
-            post(scene_syncd::api::routes::bulk_upsert_entities),
-        )
-        .route(
-            "/entities/list",
-            post(scene_syncd::api::routes::list_entities),
-        )
-        .route(
-            "/relations/bulk-upsert",
-            post(scene_syncd::api::routes::bulk_upsert_relations),
-        )
-        .route(
-            "/relations/list",
-            post(scene_syncd::api::routes::list_relations),
-        )
-        .route(
-            "/assets/upsert",
-            post(scene_syncd::api::routes::upsert_asset),
-        )
-        .route("/assets/list", post(scene_syncd::api::routes::list_assets))
-        // P6: Component, Blueprint, Realization routes
-        .route(
-            "/components/upsert",
-            post(scene_syncd::api::routes::upsert_component),
-        )
-        .route(
-            "/components/list",
-            post(scene_syncd::api::routes::list_components),
-        )
-        .route(
-            "/components/delete",
-            post(scene_syncd::api::routes::delete_component),
-        )
-        .route(
-            "/blueprints/upsert",
-            post(scene_syncd::api::routes::upsert_blueprint),
-        )
-        .route(
-            "/blueprints/list",
-            post(scene_syncd::api::routes::list_blueprints),
-        )
-        .route(
-            "/blueprints/delete",
-            post(scene_syncd::api::routes::delete_blueprint),
-        )
-        .route(
-            "/realizations/upsert",
-            post(scene_syncd::api::routes::upsert_realization),
-        )
-        .route(
-            "/realizations/list",
-            post(scene_syncd::api::routes::list_realizations),
-        )
-        .route(
-            "/realizations/update-status",
-            post(scene_syncd::api::routes::update_realization_status),
-        )
-        .route(
-            "/layouts/{scene_id}/denormalize",
-            post(scene_syncd::api::routes::denormalize_layout_route),
-        )
-        .route(
-            "/layouts/{scene_id}/nodes/{entity_id}/transform",
-            post(scene_syncd::api::routes::update_layout_node_transform),
-        )
-        .route(
-            "/layouts/{scene_id}/approve",
-            post(scene_syncd::api::routes::approve_layout),
-        )
-        .route(
-            "/layouts/{scene_id}/preview",
-            post(scene_syncd::api::routes::preview_layout_route),
-        )
-        .route(
-            "/layouts/{scene_id}/compile/preview",
-            post(scene_syncd::api::routes::compile_preview_route),
-        )
-        .route(
-            "/layouts/{scene_id}/validate",
-            post(scene_syncd::api::routes::validate_route),
-        )
-        .route(
-            "/layouts/{scene_id}/compile/plan",
-            post(scene_syncd::api::routes::compile_plan_route),
-        )
-        .route(
-            "/layouts/{scene_id}/compile/apply",
-            post(scene_syncd::api::routes::compile_apply_route),
-        )
-        .route(
-            "/realizations/{scene_id}/realize",
-            post(scene_syncd::api::routes::realize_layout_route),
-        )
-        // P10: PIE / Log Validation routes
-        .route("/unreal/pie/run", post(scene_syncd::api::routes::pie_run))
-        .route(
-            "/unreal/logs/parse",
-            post(scene_syncd::api::routes::parse_logs),
-        )
-        .route("/unreal/fix-plan", post(scene_syncd::api::routes::fix_plan))
-        .route(
-            "/procedural/create-mesh",
-            post(scene_syncd::api::routes::create_procedural_mesh_route),
-        )
-        .route(
-            "/procedural/sdf-mesh",
-            post(scene_syncd::api::routes::sdf_mesh_route),
-        )
-        .route(
-            "/procedural/superformula-mesh",
-            post(scene_syncd::api::routes::superformula_mesh_route),
-        )
-        .route(
-            "/procedural/lsystem-spline",
-            post(scene_syncd::api::routes::lsystem_spline_route),
-        )
-        .route(
-            "/procedural/wfc-grid",
-            post(scene_syncd::api::routes::wfc_grid_route),
-        )
-        // Procedural background jobs (large WFC / L-System runs)
-        .route(
-            "/procedural/jobs/submit",
-            post(scene_syncd::api::routes::procedural_job_submit_route),
-        )
-        .route(
-            "/procedural/jobs/{job_id}",
-            get(scene_syncd::api::routes::procedural_job_status_route),
-        )
-        .route(
-            "/procedural/jobs/{job_id}/cancel",
-            post(scene_syncd::api::routes::procedural_job_cancel_route),
-        )
-        .route(
-            "/procedural/jobs",
-            get(scene_syncd::api::routes::procedural_job_list_route),
-        )
+        .merge(scene_syncd::api::scene_routes::router())
+        .merge(scene_syncd::api::sync_routes::router())
+        .merge(scene_syncd::api::semantic_routes::router())
+        .merge(scene_syncd::api::layout_routes::router())
+        .merge(scene_syncd::api::pie_routes::router())
+        .merge(scene_syncd::api::procedural_routes::router())
         .layer(axum::extract::DefaultBodyLimit::max(512 * 1024 * 1024))
         .layer(TraceLayer::new_for_http())
         .layer(
